Move the REST logic to the middleware Router.
